Move insertLexicalCategory mutation into mutations.ts file
